persistSqlLabStateEnhancer and middleware
