+++ superset-frontend/src/SqlLab/utils/reduxStateToLocalStorageHelper.js
@@ -16,7 +16,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import persistState from 'redux-localstorage';
 import pick from 'lodash/pick';
 import { tableApiUtil } from 'src/hooks/apiResources/tables';
 import {
@@ -99,85 +98,24 @@ export function clearQueryEditors(queryEditors) {
   );
 }
 
-const CLEAR_ENTITY_HELPERS_MAP = {
-  tables: emptyTablePersistData,
-  queries: emptyQueryResults,
-  queryEditors: clearQueryEditors,
-  unsavedQueryEditor: qe => clearQueryEditors([qe])[0],
-};
-
-const sqlLabPersistStateConfig = {
-  paths: ['sqlLab'],
-  config: {
-    slicer: paths => state => {
-      const subset = {};
-      paths.forEach(path => {
-        // this line is used to remove old data from browser localStorage.
-        // we used to persist all redux state into localStorage, but
-        // it caused configurations passed from server-side got override.
-        // see PR 6257 for details
-        delete state[path].common; // eslint-disable-line no-param-reassign
-        if (path === 'sqlLab') {
-          subset[path] = Object.fromEntries(
-            Object.entries(state[path]).map(([key, value]) => [
-              key,
-              CLEAR_ENTITY_HELPERS_MAP[key]?.(value) ?? value,
-            ]),
-          );
-        }
-      });
-
-      const data = JSON.stringify(subset);
-      // 2 digit precision
-      const currentSize =
-        Math.round(((data.length * BYTES_PER_CHAR) / KB_STORAGE) * 100) / 100;
-      if (state.localStorageUsageInKilobytes !== currentSize) {
-        state.localStorageUsageInKilobytes = currentSize; // eslint-disable-line no-param-reassign
-      }
-
-      return subset;
-    },
-    merge: (initialState, persistedState = {}) => {
-      const result = {
-        ...initialState,
-        ...persistedState,
-        sqlLab: {
-          ...(persistedState?.sqlLab || {}),
-          // Overwrite initialState over persistedState for sqlLab
-          // since a logic in getInitialState overrides the value from persistedState
-          ...initialState.sqlLab,
-        },
-      };
-      return result;
-    },
-  },
-};
-
-export const persistSqlLabStateEnhander = persistState(
-  sqlLabPersistStateConfig.paths,
-  sqlLabPersistStateConfig.config,
-);
-
-export function rehydratePersistedState(dispatch, persistedState) {
+export function rehydratePersistedState(dispatch, state) {
   // Rehydrate server side persisted table metadata
-  persistedState.sqlLab.tables.forEach(
-    ({ name: table, schema, dbId, persistData }) => {
-      if (dbId && schema && table && persistData?.columns) {
-        dispatch(
-          tableApiUtil.upsertQueryData(
-            'tableMetadata',
-            { dbId, schema, table },
-            persistData,
-          ),
-        );
-        dispatch(
-          tableApiUtil.upsertQueryData(
-            'tableExtendedMetadata',
-            { dbId, schema, table },
-            {},
-          ),
-        );
-      }
-    },
-  );
+  state.sqlLab.tables.forEach(({ name: table, schema, dbId, persistData }) => {
+    if (dbId && schema && table && persistData?.columns) {
+      dispatch(
+        tableApiUtil.upsertQueryData(
+          'tableMetadata',
+          { dbId, schema, table },
+          persistData,
+        ),
+      );
+      dispatch(
+        tableApiUtil.upsertQueryData(
+          'tableExtendedMetadata',
+          { dbId, schema, table },
+          {},
+        ),
+      );
+    }
+  });
 }
